Server-side movement + refactoring

+++ src/config.rs
@@ -9,44 +9,44 @@ use std::io::{Read, Write};
 #[derive(Serialize, Deserialize, Debug)]
 pub struct Config {
     #[serde(default = "default_player_x")]
-    pub player_x: f32,
+    pub player_x: f64,
     #[serde(default = "default_player_y")]
-    pub player_y: f32,
+    pub player_y: f64,
     #[serde(default = "default_player_z")]
-    pub player_z: f32,
+    pub player_z: f64,
     #[serde(default = "default_mouse_speed")]
-    pub mouse_speed: f32,
+    pub mouse_speed: f64,
     #[serde(default = "default_player_speed")]
-    pub player_speed: f32,
+    pub player_speed: f64,
     #[serde(default = "default_ctrl_speedup")]
-    pub ctrl_speedup: f32,
+    pub ctrl_speedup: f64,
     #[serde(default = "default_render_distance")]
     pub render_distance: i64,
     #[serde(default = "default_tick_rate")]
     pub tick_rate: u64,
 }
 
-fn default_player_x() -> f32 {
+fn default_player_x() -> f64 {
     0.0
 }
 
-fn default_player_y() -> f32 {
+fn default_player_y() -> f64 {
     -100.0
 }
 
-fn default_player_z() -> f32 {
+fn default_player_z() -> f64 {
     0.0
 }
 
-fn default_mouse_speed() -> f32 {
+fn default_mouse_speed() -> f64 {
     0.2
 }
 
-fn default_player_speed() -> f32 {
+fn default_player_speed() -> f64 {
     5.0
 }
 
-fn default_ctrl_speedup() -> f32 {
+fn default_ctrl_speedup() -> f64 {
     15.0
 }
